% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{base_maestra_ninez}
\alias{base_maestra_ninez}
\title{Base Maestra ENDI Ronda 1}
\format{
Data frame
}
\usage{
base_maestra_ninez
}
\description{
Base integrada para análisis.
}
\keyword{datasets}
