% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_desarrollo_infantil}
\alias{dataset_desarrollo_infantil}
\title{Dataset Desarrollo Infantil ENDI Ronda 1}
\format{
Data frame
}
\usage{
dataset_desarrollo_infantil
}
\description{
Información de desarrollo infantil.
}
\keyword{datasets}
