% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_lactancia}
\alias{dataset_lactancia}
\title{Dataset Lactancia ENDI Ronda 1}
\format{
Data frame
}
\usage{
dataset_lactancia
}
\description{
Información de lactancia.
}
\keyword{datasets}
