% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_personas}
\alias{dataset_personas}
\title{Dataset Personas ENDI Ronda 1}
\format{
Data frame
}
\usage{
dataset_personas
}
\description{
Información sociodemográfica y antropométrica.
}
\keyword{datasets}
