% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dataset_salud_ninez}
\alias{dataset_salud_ninez}
\title{Dataset Salud Niñez ENDI Ronda 1}
\format{
Data frame
}
\usage{
dataset_salud_ninez
}
\description{
Información de salud infantil.
}
\keyword{datasets}
